Title: 쪼개어 배낭 채우기 구현, Time: 837ms, Memory: 52MB, Status: Runtime Error - CodeTree

+++ 250321/쪼개어 배낭 채우기 구현/implement-fractional-knapsack.java
@@ -8,12 +8,12 @@ class Jewel implements Comparable<Jewel> {
     public Jewel(int w, int v) {
         this.w = w;
         this.v = v;
-        this.p = (float) v / w;
+        this.p = (double) v / w;
     }
 
     @Override
     public int compareTo(Jewel j) {
-        return Float.compare(j.p, this.p);
+        return Double.compare(j.p, this.p);
     }
 }
